Use more OTel code for runtime attachment
